[FEAT] Written up the first version of the parser
At this stage, the parser works as expected. The
only issue here is that BibTeX format is rather
non-standard: there is no single formal
definition of this domain-specific language, and
it is hard to tell whether the format reached
the stage of maturity at any point in time.

That said, for now the parser is based the
material that I could find on the Internet and
plus some personal understanding of the format
as a user.